requirements.txt
python-dotenv>=1.0.1
numpy>=1.26.0
sentence-transformers>=2.7.0
torch>=2.1.0


